requirements.txt
asgiref==3.9.2
attrs==25.4.0
bitarray==3.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.7
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
openapi-codec==1.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
pybloom_live==4.0.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.29.0
simplejson==3.20.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xxhash==3.6.0
mosspy==1.0.9
django-schema-viewer==0.5.3
django-unfold==0.72.0
celery==5.4.0
